Added dynamic slug page
Adjusted the Product type to include the object ID that is included in MongoDB. Changed the type of the images value to be a string array instead of an any type array. Added basic CSS styling and JSX to Item component. Added the dynamic product page based on the slug property.