Started search functionality
Installed use-debounce to have prebuilt debounce function for user search input. Added query parameter to the getProducts function. Implemented the debounce, router and useEffect to reflect user input the in the URL of the shop page. Adjusted the search aggregator in productService to match results to user query, which accounts for up to 2 user typos. It matches user query against the category, brand, and name fields in the products collection.