Finished search input
Added searchparams prop to header to allow the params to be accessed by the search component in the header. Adjusted the Link components on pagination buttons to keep the search query in the domain when changing the page.  Added search prop to Search component to allow it to be stored in the text state. Wrapped the useEffect in a submit function and the input in a form to force the router to update only when user presses enter on the input.