Added react-query
Installed react-query to allow for a global search component. Implemented the QueryClientProvider from react-query. Removed search params from all pages to prepare to convert to react-query. Created ReactQueryClientProvider component to wrap the layout page and its children in the QueryClientProvider.  Removed all react hooks except for useState from search component to prepare for useQuery addition. Added another function to productService to allow the search component to take in text input and retrieve products that match the name field in the products database.