Started slug page design
Separated the navbar/sidebar on home page into its own client-side component so that the home page can be server-side. Changed the links on home page to redirect to the search page with the query searchParams set to the link's category. Created ImageSet component for the left side of the slug page. Created CSS styling for the slug page, ImageSet, and HomeNav components.