Gave the redux provider to layout
Wrapped children in layout.tsx with the redux Provider component that has the global store passed to it as a prop. Created custom useAppDispatch and useAppSelector constants to apply types to useDispatch and useSelector hooks respectively. Applied useAppDispatch to the CartInfo component and useAppSelector to the header that displays the shopping cart.