Implemented cart quantity
Took the useAppSelector custom hook and used it to retrieve the total amount of items in the cart and display them in a span next to the cart icon. Reduced debounce input time on search input to 600ms. Added css styling for the item amount span tag for all breakpoints. Fixed an alignment issue on the slug page for the h2.