Implemented cart page
Took the main shopping cart and made it into a client-side component called CartTable. Changes rendered image based on if color is selected or not. Uses useAppSelector and useAppDispatch custom hooks to to map the cart items and remove the cart items respectively via clearCart and removeFromCart reducers. Changed the render of cart amount on the cart svg so that it renders null if there are no items in the cart. Made default state of color and size in itemValues object "N/A" so that is shows something in the cart instead of just an empty string when color or size aren't applicable. Started CSS for the cart page and the table component.