Implemented redux persist
Figured out how to add the noop storage and create a persistent global store. Made createNoopStorage function that defaults to using keys and values of any type, although the persistConfig is set to use the root key. The noop storage is used if there is no web storage to create. The state is set to persist no matter if the page is exited or refreshed. The ReduxProvider component now has a PersistGate with the store passed to it via the persistor prop that uses the persistStore function from redux-persist. Added priority prop to the Image on the Item component that appears on the shop page. Made some minor CSS adjustments to the footer and shop page.