Started user authentication
Installed axios to post user data to a free API called RealWorld that will store the data so that it can be retrieved for authentication later. Created authSlice for the user authentication and passed its reducer to the store. Created login and sign up pages. Created SignInForm component for the state of the form and to dispatch the authSlice builder, which will return different things based on if the user is authenticated or not.