Created login and signup pages
Made login and signup pages for users to access or create their accounts. Made SignInForm and SignUpForm components that will be used to pull user data from the realWorld API and authenticate it or push it up to the API.  Renamed component folder to auth and CSS file to auth to make it reusable with both components. Basic CSS for the form components and made them centered on the page.