Started user authentication
Added user authentication to the SignInForm and SignUpForm components. On successful login, the user will be redirected to the shop page and a token will be added to local storage that is unique to the user. The login will do the same and pass the email and password to to the users/login endpoint of the RealWorld API, while the sign up will take email, password, and username and send it to the users endpoint. Added the login cases to the builder in extraReducers - they do the same thing as the sign up cases. Changed the join route to blogs.