Got user authentication working
Made login and register functions that will pass the user to the shop page on successful login or account creation or display an error message if unsuccessful. Created a reducer in authSlice that sets the currentUser to undefined on sign out. Made a change to header that will have a sign out button if there is a currentUser in global state or show the sign in link if not.