Created user menu UI
Created CSS and a React useState that will toggle CSS classes on an unordered list for the user when there is one logged in. The list has a account page link and a sign out button. The account page will take the username, password, and email that is set to currentUser in global state and display them.