Started account page
Fixed an unexpected auth bug by using the unwrap method chained on the dispatch calls to the login and register thunks. Made AccountInfo component for the account page that will access the auth state using the custom useAppSelector hook. Adjusted the type of currentUser in auth to not have an undefined value in order to set image property for the image that is passed when user logs in. Started CSS for the account page and the AccountInfo component.