Implemented contact form functionality
Created API route on project to allow for messages on contact form to be passed into the contacts collection in the MongoDB. Returns messages based on if the message was successfully passed or not. Updated metadata on contact page. Implemented state on the contact form to pass each field into the post request in the fetch API. Made contact schema and model for the new collection in the database. Removed unused user model and schema. Made CSS for the error/success message on contact page.