Created dynamic blog route
Created a dynamic blog page for each blog that is in the database. Made a blogs collection in the MongoDB cluster. Added metadata to a couple pages. Blogs page maps out all the blogs in the database, with each containing a Blog component that links to the blog page. Created a blog service that will retrieve the blogs from the database. Contains two methods: one to get all blogs and one to find a specific blog when the user clicks on a blog post on the blogs page. Created schema, model, and type for the blog posts. Made router redirect to shop page when the signOut reducer is activated. Created CSS for blog component and blogs page.