Created blog posts
Created blog posts in the MongoDB collection. Displayed the blog post paragraphs on the dynamic page, as well as the title, author, and date of the posts. Added a key for each Blog component on the blogs page. Made CSS for the blog posts so that they are responsive at each resolution.