Final commit before production
Added HTML entities for various escape characters across multiple files. Added priority prop to image on blog posts and set the src to be the image property of the post in the database. Set the account page profile picture to be the site logo when user is not logged in to stop an error. Fixed bad links on home page. Added router to dependency array of useEffect in Search component.